Update WebToolBarUI.java
WebToolBarUI sets the standard layout of a weblaf toolbar, ToolbarLayout, in its installUI method. 
But if someone changes the layout after that, WebToolBarUI should respect that and don't make any changes.
As an example I have subclassed a JToolbar, setting a GridLayout in its constructor, and adding some buttons in two rows to the toolbar.
Because I ned to call the toolbars setFloatable method,  the actual code would change my layout, with the above changes it wouldn't.