add 'noshade' option to label.
although the default skin does not add a shade to the regular label,
other skin might do. then for an application in order to declare a
'plain' label without shade, it should be possible to call
`putClientProperty("styleId", "noshade")`. this will currently
throw an exception with the default skin because no such style
exists.