Define disabled icon for WebCollapsiblePane
For this to work correctly, the setEnabled
method needs to be overriden and needs to
update the enabled state of the expanded button.

Currently the icon generated by createDisabledCopy
is not so much different than the enabled one,
as the opacity should be reduced even more in
ImageUtils (later maybe).