Revamped `WebBreadcrumb` component and added new shape detection settings to make mouse events effective area more precise.
New `WebBreadcrumb` component is now fully styleable and includes a few new settings you might find useful. It also doesn't use any custom components like `WebBreadcrumbLabel` or `WebBreadcrumbButton` anymore - it dynamically adjusts added components style (if that feature is enabled) to fit into `WebBreadcrumb`. Data like the element's progress is now stored separately within `WebBreadcrumb` itself and only exists as long as the component stays on the `WebBreadcrumb`. As for the default style I went with two options for starting (in each of the skins) - both have similar color scheme, but use a bit different components layout.

Shape detection is a new setting that I did plan to add somewhere in-between other major changes, but never really got to do it. Although it was quite necessary for the `WebBreadcrumb` rework due to the way its elements are laid out and used. So this setting does two things - it makes all mouse events (for components with this feature enabled) to take shape into account. That means that if you, for instance, have a button with a large shadow and/or margin - it will not react to mouse clicks outside of the visible button border bounds. Second thing - this change surely does a big hit on performance of all `contains()` method calls, but in the grand scheme of things it is completely irrelevant and doesn't affect neither UI nor mouse events performance while adding quite a bit of precision to your UI elements. If at any point you would feel that this feature is negatively affecting your UI performance - you can easily disable it through `WebLookAndFeel.setShapeDetectionEnabled(false)` call.

Breadcrumb
- WebBreadcrumb.java - Reworked and now fully supports styling and multiple new components as breadcrumb elements
- BreadcrumbElementData.java - Class containing information about all elements available in the breadcrumb (only supported ones)
- BreadcrumbLabelPainter.java - `BreadcrumbElementPainter` implementation for `JLabel` component
- BreadcrumbStyledLabelPainter.java - `BreadcrumbElementPainter` implementation for `WebStyledLabel` component
- BreadcrumbButtonPainter.java - `BreadcrumbElementPainter` implementation for `JButton` component
- DecorationState.java, BreadcrumbElementUtils.java - Added custom breadcrumb element states and their usage
- breadcrumb.xml - Added support for all different child components
- breadcrumb.xml - Added new base styles
- breadcrumb.xml - Added dark skin support

Styling
- Added support for `contains(x,y)` methods in UI, Painter and Decoration implementations
- WebLookAndFeel.java - Added global `shapeDetectionEnabled` setting (`true` by default) to allow mouse events "see" component shapes
- Added methods to enable or disable shape detection setting (per component instance) into all styleable components

Decoration
- IconTextLayout.java - Fixed text positioning for `TRAILING` alignment with RTL orientation
- AbstractClipBackground.java - New abstract `IBackground` implementation that provides an API to clip other `IBackground` implementations
- MovingHighlightBackground.java - Added proper clipping with provided shape to avoid overextending background beyond that shape
- Sides.java, SidesConverter.java - Added "none" constant for convenience, it can be specified as value

SplitButton
- SplitButtonPainter.java - Fixed RTL arrow button positioning
- SplitButtonPainter.java - Fixed possible NPE and refactored painting code a bit
- WebSplitButton.java, SplitButtonPainter.java - Renamed `splitIcon` into `menuIcon` for convenience and adjusted all code accordingly

ProgressBar
- progressbar.xml - Tuned down background highlight pass frequency for `intermediate` progress state

ToolbarSeparator
- WebToolBarSeparator.java - Overridden `setSeparatorSize` method to avoid calling `super.updateUI()` which caused issues when WebLaF is not installed

ComboBox
- WebComboBoxUI.java - Replaced custom `JList` with a custom `WebList` usage to avoid issues when WebLaF is not installed

FileChooser
- filechooser.xml - Changed scrollbar style for the list/table area to improve experience

FileList
- filelist.xml - Added extra styles for `filelist-tiles` and `filelist-icon` base styles

ButtonPopup
- WebButtonPopup.java - Marked as deprecated, will be reworked in future before removal

Layout
- VerticalFlowLayout.java - Fixed RTL elements positioning
- HorizontalFlowLayout.java - Minor code refactoring

Utilities
- ComponentHighlighter.java - Simplified margin and padding retrieval and fixed issue with RTL case display
- PainterSupport.java - Added methods to retrieve margin and padding that account for component's orientation
- ContainerMethods.java - Added method for iterating through the container's children, will mostly be useful for JDK8+ users
- ContainerMethods.java, ContainerMethodsImpl.java - Refactored JavaDoc and fixed some typos
- UnselectableButtonGroup.java - Added method to group all `AbstractButton`s from a single `Container`, minor code refactoring
- TableRowHeightOptimizer.java - Fixed exceptions that could occur on smaller tables, made this utility orientation-aware

Demo
- WebBreadcrumbExample.java - New example demonstrating revamped `WebBreadcrumb` component in action
- MiscGroup.java, ShapeDetectionExample.java - New example demonstrating new WebLaF shape detection feature
- Minor examples and groups code refactoring